connect to stores

+++ EtradeSystem/src/Domain/Users/ExternalService/Payment/PaymentAdaptee.java
@@ -4,10 +4,10 @@ package Domain.Users.ExternalService.Payment;
 import java.time.LocalTime;
 
 public class PaymentAdaptee {
-    public boolean payment(int cardNum, LocalTime expDate, int cvv, int price,int cardTo){
+    public boolean payment(int cardNum, LocalTime expDate, int cvv, double price,int cardTo){
         return false;
     }
-    public boolean canPay(int cardFrom, LocalTime expDate, int cvv,int price){
+    public boolean canPay(int cardFrom, LocalTime expDate, int cvv,double price){
         return false;
     }
 }
